+++ src/converter/data/dataConverter.ts
@@ -55,9 +55,12 @@ import {
     DataRepresentationTypeEnum,
 } from "../data/dataRepresentationScale";
 
-import { createVarianceConverter } from "../variance/varianceConverter";
+import { createVarianceConverterByType } from "../variance";
 
-import { DataFormatter } from "../data/dataFormatter";
+import {
+    getFormattedDate,
+    getFormattedValueWithFallback,
+} from "../data/dataFormatter";
 
 export interface IColumnGroup {
     name: string;
@@ -320,7 +323,10 @@ export class DataConverter implements IConverter<IDataConverterOptions, IDataRep
                     dataRepresentation.sortedSeries.push(series);
                 }
 
-                const y: number = valueColumnGroup.values[columnIndex] || 0;
+                const y: number = this.parseValue(
+                    valueColumnGroup.values[columnIndex],
+                    dataRepresentation.series[columnIndex].settings.values.treatEmptyValuesAsZero,
+                );
 
                 const dataPoint: IDataRepresentationPoint = {
                     index: dataRepresentation.series[columnIndex].points.length,
@@ -343,25 +349,27 @@ export class DataConverter implements IConverter<IDataConverterOptions, IDataRep
                     x,
                 );
 
-                dataRepresentation.series[columnIndex].y.min = this.getMin(
-                    dataRepresentation.series[columnIndex].y.min,
-                    y,
-                );
+                if (!isNaN(y)) {
+                    dataRepresentation.series[columnIndex].y.min = this.getMin(
+                        dataRepresentation.series[columnIndex].y.min,
+                        y,
+                    );
 
-                dataRepresentation.series[columnIndex].y.max = this.getMax(
-                    dataRepresentation.series[columnIndex].y.max,
-                    y,
-                );
+                    dataRepresentation.series[columnIndex].y.max = this.getMax(
+                        dataRepresentation.series[columnIndex].y.max,
+                        y,
+                    );
 
-                dataRepresentation.series[columnIndex].ySparkline.min = this.getMin(
-                    dataRepresentation.series[columnIndex].ySparkline.min,
-                    y,
-                );
+                    dataRepresentation.series[columnIndex].ySparkline.min = this.getMin(
+                        dataRepresentation.series[columnIndex].ySparkline.min,
+                        y,
+                    );
 
-                dataRepresentation.series[columnIndex].ySparkline.max = this.getMax(
-                    dataRepresentation.series[columnIndex].ySparkline.max,
-                    y,
-                );
+                    dataRepresentation.series[columnIndex].ySparkline.max = this.getMax(
+                        dataRepresentation.series[columnIndex].ySparkline.max,
+                        y,
+                    );
+                }
 
                 const tooltip: string = tooltipColumnGroup
                     && tooltipColumnGroup.values
@@ -392,7 +400,7 @@ export class DataConverter implements IConverter<IDataConverterOptions, IDataRep
     }
 
     private postProcessData(dataRepresentation: IDataRepresentation, settings: Settings): void {
-        dataRepresentation.series.forEach((series: IDataRepresentationSeries, seriesIndex: number) => {
+        dataRepresentation.series.forEach((series: IDataRepresentationSeries) => {
             series.x.initialMin = series.x.min;
             series.x.initialMax = series.x.max;
 
@@ -424,14 +432,14 @@ export class DataConverter implements IConverter<IDataConverterOptions, IDataRep
 
             const endDataPoint: IDataRepresentationPoint = series.points[series.points.length - 1];
 
-            series.variance = createVarianceConverter()
+            series.variance = createVarianceConverterByType(series.settings.variance.shouldCalculateDifference)
                 .convert({
                     firstDataPoint: startDataPoint,
                     secondDataPoint: endDataPoint,
                 });
 
-            series.formattedDate = DataFormatter.getFormattedDate(startDataPoint.x, settings.date.getFormat());
-            series.formattedVariance = DataFormatter.getFormattedVariance(series.variance);
+            series.formattedDate = getFormattedDate(startDataPoint.x, settings.date.getFormat());
+            series.formattedVariance = getFormattedValueWithFallback(series.variance, series.settings.variance);
 
             series.dateDifference = this.getDaysBetween(endDataPoint.x, startDataPoint.x);
 
@@ -506,22 +514,52 @@ export class DataConverter implements IConverter<IDataConverterOptions, IDataRep
     }
 
     private getMin<Type>(originalValue: Type, value: Type): Type {
-        if (originalValue === undefined || originalValue === null) {
+        const isOriginalValueValid: boolean = this.isValueValid(originalValue);
+        const isValueValid: boolean = this.isValueValid(value);
+
+        if (isOriginalValueValid && isValueValid) {
+            return value < originalValue
+                ? value
+                : originalValue;
+        }
+
+        if (isValueValid) {
             return value;
         }
 
-        return value < originalValue
-            ? value
-            : originalValue;
+        return undefined;
     }
 
     private getMax<Type>(originalValue: Type, value: Type): Type {
-        if (originalValue === undefined || originalValue === null) {
+        const isOriginalValueValid: boolean = this.isValueValid(originalValue);
+        const isValueValid: boolean = this.isValueValid(value);
+
+        if (isOriginalValueValid && isValueValid) {
+            return value > originalValue
+                ? value
+                : originalValue;
+        }
+
+        if (isValueValid) {
             return value;
         }
 
-        return value > originalValue
-            ? value
-            : originalValue;
+        return undefined;
+    }
+
+    private isValueValid<Type>(value: Type): boolean {
+        return value != null;
+    }
+
+    private parseValue(value: number, treatEmptyValuesAsZero: boolean): number {
+        if (isFinite(value) && value != null) {
+            return value;
+        }
+
+        if (treatEmptyValuesAsZero) {
+            return 0;
+        }
+
+        return NaN;
     }
 }