+++ src/visualComponent/mainChart/axisComponent.ts
@@ -66,7 +66,7 @@ export class AxisComponent extends BaseComponent<IVisualComponentConstructorOpti
     public render(options: IAxisComponentRenderOptions): void {
         const { settings } = options;
 
-        if (settings.shouldBeShown()) {
+        if (settings.shouldBeShown) {
             this.show();
             this.renderComponent(options);
         } else {