+++ src/visualComponent/mainChart/hoverLabelComponent.ts
@@ -26,22 +26,28 @@
 
 import { valueFormatter } from "powerbi-visuals-utils-formattingutils";
 
-import { ChartLabelBaseComponent } from "./chartLabelBaseComponent";
+import { ChartLabelBaseComponent, IRenderGroup } from "./chartLabelBaseComponent";
 
-import { createVarianceConverter } from "../../converter/variance/varianceConverter";
+import { createVarianceConverterByType } from "../../converter/variance";
 
-import { DataFormatter } from "../../converter/data/dataFormatter";
 import { IDataRepresentationPoint } from "../../converter/data/dataRepresentation";
 
-import { VarianceChecker } from "../../converter/variance/varianceChecker";
+import {
+    getFormattedDate,
+    getFormattedValueWithFallback,
+} from "../../converter/data/dataFormatter";
+
+import { isValueValid } from "../../utils/valueUtils";
 
 import { IVerticalReferenceLineComponentRenderOptions } from "../verticalReferenceLineComponent";
 import { IVisualComponentConstructorOptions } from "../visualComponentConstructorOptions";
 
 import { KpiOnHoverDescriptor } from "../../settings/descriptors/kpi/kpiOnHoverDescriptor";
+import { VarianceDescriptor } from "../../settings/descriptors/varianceDescriptor";
 
 export interface IHoverLabelComponentRenderOptions extends IVerticalReferenceLineComponentRenderOptions {
     kpiOnHoverSettings: KpiOnHoverDescriptor;
+    varianceSettings: VarianceDescriptor;
 }
 
 export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComponentRenderOptions> {
@@ -60,6 +66,7 @@ export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComp
             dataPoint,
             kpiOnHoverSettings,
             dateSettings,
+            varianceSettings,
         } = options;
 
         this.updateFormatting(this.element, kpiOnHoverSettings);
@@ -74,16 +81,11 @@ export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComp
 
         const latestDataPoint: IDataRepresentationPoint = series.points[series.points.length - 1];
 
-        const variance: number = createVarianceConverter()
+        const variance: number = createVarianceConverterByType(varianceSettings.shouldCalculateDifference)
             .convert({
                 firstDataPoint: dataPoint,
                 secondDataPoint: latestDataPoint,
             });
-
-        const formatter: valueFormatter.IValueFormatter = DataFormatter.getValueFormatter(
-            latestDataPoint.y,
-            series.settings.values,
-        );
 
         this.renderGroup(
             this.headerSelector,
@@ -98,14 +100,14 @@ export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComp
 
         );
 
-        const isVarianceValid: boolean = VarianceChecker.isVarianceValid(variance);
+        const isVarianceValid: boolean = isValueValid(variance);
 
         this.renderGroup(
             this.bodySelector,
             [
                 {
                     color: kpiOnHoverSettings.valueColor,
-                    data: formatter.format(latestDataPoint.y),
+                    data: getFormattedValueWithFallback(latestDataPoint.y, series.settings.values),
                     fontSizeInPt: kpiOnHoverSettings.valueFontSize,
                     isShown: kpiOnHoverSettings.isValueShown,
                 },
@@ -113,7 +115,7 @@ export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComp
                     color: isVarianceValid
                         ? kpiOnHoverSettings.varianceColor
                         : kpiOnHoverSettings.varianceNotAvailableColor,
-                    data: `(${DataFormatter.getFormattedVariance(variance)})`,
+                    data: `(${getFormattedValueWithFallback(variance, varianceSettings)})`,
                     fontSizeInPt: isVarianceValid
                         ? kpiOnHoverSettings.varianceFontSize
                         : kpiOnHoverSettings.varianceNotAvailableFontSize,
@@ -125,22 +127,28 @@ export class HoverLabelComponent extends ChartLabelBaseComponent<IHoverLabelComp
             ],
         );
 
+        const dateGroup: IRenderGroup[] = [
+            {
+                color: kpiOnHoverSettings.currentValueColor,
+                data: getFormattedValueWithFallback(dataPoint.y, series.settings.values),
+                fontSizeInPt: kpiOnHoverSettings.currentValueFontSize,
+                isShown: kpiOnHoverSettings.isCurrentValueShown,
+            },
+            {
+                color: kpiOnHoverSettings.dateColor,
+                data: getFormattedDate(dataPoint.x, dateSettings.getFormat()),
+                fontSizeInPt: kpiOnHoverSettings.dateFontSize,
+                isShown: kpiOnHoverSettings.isDateShown,
+            },
+        ];
+
+        const dateGroupToRender: IRenderGroup[] = kpiOnHoverSettings.isCurrentValueLeftAligned
+            ? dateGroup
+            : [...dateGroup].reverse(); // Makes dateGroupBase array immutable
+
         this.renderGroup(
             this.footerSelector,
-            [
-                {
-                    color: kpiOnHoverSettings.currentValueColor,
-                    data: formatter.format(dataPoint.y),
-                    fontSizeInPt: kpiOnHoverSettings.currentValueFontSize,
-                    isShown: kpiOnHoverSettings.isCurrentValueShown,
-                },
-                {
-                    color: kpiOnHoverSettings.dateColor,
-                    data: DataFormatter.getFormattedDate(dataPoint.x, dateSettings.getFormat()),
-                    fontSizeInPt: kpiOnHoverSettings.dateFontSize,
-                    isShown: kpiOnHoverSettings.isDateShown,
-                },
-            ],
+            dateGroupToRender,
         );
     }
 }