+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 2.1.0
+* Adds Decimal Places, Precision and Format for variance values
+* Fixes auto font size issue
+* Adds Stale Data options (title pattern, threshold, color, and background)
+* Adds alignment for subtitle
+* Adds "On Hover Current Value" alignment
+* Allows calculating a difference instead of growth for percentage metrics
+* Allows to treat empty values as zero
+
 ## 2.0.0
 * API 2.2.0
 * PBIVIZ 3.x.x